clock: move timer functions to libtrx
