--- src/libtrx/game/clock.c
@@ -1,15 +1,0 @@
-#include "game/clock.h"
-
-#include <stdio.h>
-#include <time.h>
-
-size_t Clock_GetDateTime(char *const buffer, const size_t size)
-{
-    time_t lt = time(0);
-    struct tm *tptr = localtime(&lt);
-
-    return snprintf(
-        buffer, size, "%04d%02d%02d_%02d%02d%02d", tptr->tm_year + 1900,
-        tptr->tm_mon + 1, tptr->tm_mday, tptr->tm_hour, tptr->tm_min,
-        tptr->tm_sec);
-}
