+++ src/libtrx/game/clock/common.c
@@ -1,10 +1,10 @@
-#include "config.h"
 #include "game/clock.h"
 
 #include <SDL2/SDL_stdinc.h>
 #include <SDL2/SDL_timer.h>
 #include <stdbool.h>
 #include <stdint.h>
+#include <time.h>
 
 static Uint64 m_LastCounter = 0;
 static Uint64 m_InitCounter = 0;
@@ -18,6 +18,32 @@ void Clock_Init(void)
     m_Counter = SDL_GetPerformanceCounter();
 }
 
+size_t Clock_GetDateTime(char *const buffer, const size_t size)
+{
+    time_t lt = time(0);
+    struct tm *tptr = localtime(&lt);
+
+    return snprintf(
+        buffer, size, "%04d%02d%02d_%02d%02d%02d", tptr->tm_year + 1900,
+        tptr->tm_mon + 1, tptr->tm_mday, tptr->tm_hour, tptr->tm_min,
+        tptr->tm_sec);
+}
+
+int32_t Clock_GetFrameAdvance(void)
+{
+    return Clock_GetCurrentFPS() == 30 ? 2 : 1;
+}
+
+int32_t Clock_GetLogicalFrame(void)
+{
+    return Clock_GetHighPrecisionCounter() * LOGIC_FPS / 1000.0;
+}
+
+int32_t Clock_GetDrawFrame(void)
+{
+    return Clock_GetHighPrecisionCounter() * Clock_GetCurrentFPS() / 1000.0;
+}
+
 double Clock_GetHighPrecisionCounter(void)
 {
     return (SDL_GetPerformanceCounter() - m_InitCounter) * 1000.0
@@ -27,7 +53,7 @@ double Clock_GetHighPrecisionCounter(void)
 int32_t Clock_SyncTicks(void)
 {
     m_LastCounter = m_Counter;
-    const double fps = g_Config.rendering.fps;
+    const double fps = Clock_GetCurrentFPS();
 
     const double frequency = (double)m_Frequency / Clock_GetSpeedMultiplier();
     const Uint64 target_counter = m_LastCounter + (frequency / fps);
