+++ src/libtrx/include/libtrx/game/clock.h
@@ -1,6 +1,4 @@
 #pragma once
 
-#include <stddef.h>
-
-size_t Clock_GetDateTime(char *buffer, size_t size);
-extern double Clock_GetHighPrecisionCounter(void);
+#include "clock/common.h"
+#include "clock/timer.h"
