+++ src/libtrx/meson.build
@@ -68,7 +68,8 @@ sources = [
   'event_manager.c',
   'filesystem.c',
   'game/backpack.c',
-  'game/clock.c',
+  'game/clock/common.c',
+  'game/clock/timer.c',
   'game/console/cmd/config.c',
   'game/console/cmd/die.c',
   'game/console/cmd/end_level.c',
