+++ src/tr1/game/clock.c
@@ -11,41 +11,6 @@
 #include <math.h>
 #include <stdio.h>
 
-static double M_GetElapsedUnit(CLOCK_TIMER *const timer, const double unit);
-static bool M_CheckElapsedUnit(
-    CLOCK_TIMER *const timer, const double how_often, const double unit,
-    bool bypass_turbo_cheat);
-
-static double M_GetElapsedUnit(CLOCK_TIMER *const timer, const double unit)
-{
-    assert(timer != NULL);
-    const double delta = Clock_GetHighPrecisionCounter() - timer->prev_counter;
-    const double multiplier = Clock_GetSpeedMultiplier() / 1000.0;
-    const double frames = delta * multiplier * unit;
-    Clock_ResetTimer(timer);
-    return frames;
-}
-
-static bool M_CheckElapsedUnit(
-    CLOCK_TIMER *const timer, const double how_often, const double unit,
-    bool bypass_turbo_cheat)
-{
-    assert(timer != NULL);
-    const double delta = Clock_GetHighPrecisionCounter() - timer->prev_counter;
-    const double multiplier =
-        (bypass_turbo_cheat ? 1.0 : Clock_GetSpeedMultiplier()) / 1000.0;
-    const double frames = delta * multiplier * unit;
-    if (g_Config.rendering.fps != timer->prev_fps) {
-        Clock_ResetTimer(timer);
-        return false;
-    }
-    if (frames >= how_often) {
-        Clock_ResetTimer(timer);
-        return true;
-    }
-    return false;
-}
-
 int32_t Clock_GetTurboSpeed(void)
 {
     return g_Config.rendering.turbo_speed;
@@ -75,63 +40,7 @@ double Clock_GetSpeedMultiplier(void)
     }
 }
 
-int32_t Clock_GetFrameAdvance(void)
+int32_t Clock_GetCurrentFPS(void)
 {
-    return g_Config.rendering.fps == 30 ? 2 : 1;
-}
-
-int32_t Clock_GetLogicalFrame(void)
-{
-    return Clock_GetHighPrecisionCounter() * LOGIC_FPS / 1000.0;
-}
-
-int32_t Clock_GetDrawFrame(void)
-{
-    return Clock_GetHighPrecisionCounter() * g_Config.rendering.fps / 1000.0;
-}
-
-void Clock_ResetTimer(CLOCK_TIMER *const timer)
-{
-    assert(timer != NULL);
-    timer->prev_counter = Clock_GetHighPrecisionCounter();
-    timer->prev_fps = g_Config.rendering.fps;
-}
-
-double Clock_GetElapsedLogicalFrames(CLOCK_TIMER *const timer)
-{
-    return M_GetElapsedUnit(timer, LOGIC_FPS);
-}
-
-double Clock_GetElapsedDrawFrames(CLOCK_TIMER *const timer)
-{
-    return M_GetElapsedUnit(timer, g_Config.rendering.fps);
-}
-
-double Clock_GetElapsedMilliseconds(CLOCK_TIMER *const timer)
-{
-    return M_GetElapsedUnit(timer, 1000.0);
-}
-
-bool Clock_CheckElapsedLogicalFrames(
-    CLOCK_TIMER *const timer, const int32_t how_often)
-{
-    return M_CheckElapsedUnit(timer, how_often, LOGIC_FPS, false);
-}
-
-bool Clock_CheckElapsedDrawFrames(
-    CLOCK_TIMER *const timer, const int32_t how_often)
-{
-    return M_CheckElapsedUnit(timer, how_often, g_Config.rendering.fps, false);
-}
-
-bool Clock_CheckElapsedMilliseconds(
-    CLOCK_TIMER *const timer, const int32_t how_often)
-{
-    return M_CheckElapsedUnit(timer, how_often, 1000, false);
-}
-
-bool Clock_CheckElapsedRawMilliseconds(
-    CLOCK_TIMER *const timer, const int32_t how_often)
-{
-    return M_CheckElapsedUnit(timer, how_often, 1000, true);
+    return g_Config.rendering.fps;
 }
