+++ src/tr2/game/clock.h
@@ -1,6 +1,3 @@
 #pragma once
 
-#include <stdint.h>
-
-double Clock_GetHighPrecisionCounter(void);
-int32_t Clock_GetLogicalFrame(void);
+#include <libtrx/game/clock.h>
